Split out initializer function
Adds an index.js file with lib/sphero.js initializer - export of that
file is now the Sphero class for better testing.

Also adds a check to this instance of Sphero to check that it's being
called as a constructor (e.g. with the `new` keyword), rather than as
a normal function.